requirements.txt
# Core dependencies
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# AI/ML
langchain==0.1.0
langgraph==0.0.26
langsmith==0.0.77
openai==1.6.1
anthropic==0.8.1

# Memory system
mem0ai==0.0.15

# Blockchain
cdp-sdk==0.0.5
web3==6.15.0

# Database
google-cloud-firestore==2.14.0
google-cloud-bigquery==3.13.0
google-cloud-secret-manager==2.17.0

# API
fastapi==0.108.0
uvicorn[standard]==0.25.0
websockets==12.0
httpx==0.26.0

# Monitoring
prometheus-client==0.19.0
google-cloud-monitoring==2.17.0
google-cloud-logging==3.8.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
pytest-mock==3.12.0

# Development
black==23.12.1
flake8==7.0.0
mypy==1.7.1
ipython==8.19.0

# Utilities
aiohttp==3.9.1
asyncio==3.4.3
rich==13.7.0
click==8.1.7